:recycle: refactor: refactor the globalSelector to settingsSelector
